api/requirements.txt
boto3==1.34.98
python-dotenv==1.0.1
qrcode==7.4.2
PyJWT==2.8.0
